requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
PyYAML==6.0.1
kivy==2.1.0
kivymd==1.1.1
plyer==2.1.0
